# docs/swagger.yaml
definitions:
  content.Content:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      version:
        type: integer
    type: object
  response.ApiResponse:
    properties:
      message: {}
      status:
        type: string
    type: object
  response.ApiResponseSuccess:
    properties:
      data: {}
      status:
        type: string
    type: object
  user.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      photo:
        type: string
      role:
        type: string
      total_point:
        type: integer
    type: object
info:
  contact: {}
  description: Berikut API yang digunakan untuk mini project
  title: Clean Hexa Sample API
  version: "1.0"
paths:
  /akun/register:
    post:
      consumes:
      - application/json
      description: create user adn save to db
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Create user
      tags:
      - accounts
  /users/getall:
    get:
      consumes:
      - application/json
      description: Retrieve list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "403":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Get all users
      tags:
      - accounts
  /v1/content:
    post:
      consumes:
      - application/json
      description: create content with data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.Content'
      summary: Create content
      tags:
      - content
swagger: "2.0"